pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-cost-env"
version = "1.0.0"
description = "OpenEnv-compatible AWS cost optimization environment with typed client and server."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "openai>=1.30.1",
    "pydantic>=2.7.0",
    "requests>=2.32.3",
    "uvicorn>=0.30.0",
    "openenv-core>=0.2.0",
    "websocket-client>=1.8.0",
    "websockets>=12.0",
]

[project.scripts]
server = "server.run:main"

[tool.setuptools]
packages = ["aws_cost_env", "server"]
